requirements.txt
Pillow>=10.0.0
mss>=9.0.0

